proto implimentation of start and events
there is a start screen laid out
the event system is MORE in place but it still needs to be connected to things. right now there is a ui script involved and the general game manager can do more. it's a bit of a mess but there's a lot of it on paper.

Very close to having this be playable with ui sin landmarks. 

guess I should add those too before reallyyyyy getting into things lol.