event system works
the manager checks to see if our car has active events, if it does, it takes away the turn button. 
the event panels are created by the car via the ui manager, when they are closed, they tell the car. the car is checking to see when there are not event panels.
The manager is checking the car to see if it has no active events via a bool. when this happens, the manager asks the ui manager to turn the turn button on again.