I am scared and tipsy
why AM I coding liek this? Iw as playing jackbox with the boys and wanted to push because I was coding on the side in breaks
Uhhh I added in sick checks in the morning, changed batteries to one group of charge, I think that's it?

adding in the rest of the sick checks, the settler should know how to assign time to itself in that way each turn