actual death
when a settler dies, there is a function in the CAR class to remove them from the stat block.

I also had to tell the sick checks to make sure the settler isn't dead.

So a big issue right now is that settlers are not removed from the list when they die (it messes with UI calculations reeaalllyyy bad) so since we don't do too much individual checking of settlers, we just go ahead and poitn out when we don't want the CAR to include dead settlers

It's a prototype, this is bad, I know, but it works.