checkpoints now totally work and have events!
BIG NOTE: the first node in our chanin is cosmetic
the ACTUAL first node in the list is node0. I don not properly understand how I laid it out but the first node in a chain should always be the one the car goes to first.
Otherwise it's all nice and tighty.

Oh and for the sake of turning off the turn button, make sure you're clear about when events are active and not active in the CAR class. using the isActive bool and the numActive int.